Add Favorites page
Adds Favorites page and functionality to mark a project as favorite.
Uses the same tabel for projects, currencies and favorites.
A lot of refactoring to remove the duplication.
