+++ app/src/components/Footer.css
@@ -28,3 +28,7 @@
   margin-top: 1.5em;
   margin-bottom: 0.5em;
 }
+
+.cashflow-indev-message {
+  margin-bottom: 16px;
+}
