+++ app/src/components/ProjectsNavigation.css
@@ -1,20 +1,20 @@
-.cashflow-head {
+.projects-navigation {
   display: flex;
   align-items: center;
   margin: 1em 0;
   padding: 0 16px;
 }
 
-.cashflow-head span {
+.projects-navigation span {
   margin: 18px 0 0 16px;
   padding-left: 16px;
 }
 
-.cashflow-head > :nth-child(2) {
+.projects-navigation > :nth-child(2) {
   border-left: 1px solid #3d4450;
 }
 
-.cashflow-head h1 {
+.projects-navigation h1 {
   font-family: 'Source Sans Pro', sans-serif;
   margin: 0;
   color: rgba(0, 0, 0, 0.7) !important;
