+++ app/src/pages/CashflowMobile.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import ProjectsMobile from './Projects/ProjectsMobile'
-import withProjectsData from './Projects/withProjectsData'
+import withProjectsDataMobile from './Projects/withProjectsDataMobile'
 
 const CashflowMobile = props => <ProjectsMobile type={'erc20'} {...props} />
 
-export default withProjectsData({type: 'erc20'})(CashflowMobile)
+export default withProjectsDataMobile({type: 'erc20'})(CashflowMobile)
