+++ app/src/pages/Detailed/DetailedHeader.css
@@ -32,6 +32,16 @@
   margin-top: 5px;
 }
 
+.detailed-favorite {
+  margin-top: 14px;
+  margin-left: 15px;
+  cursor: pointer;
+}
+
+.detailed-favorite i {
+  margin-right: 0;
+}
+
 .pending-h1 {
   height: 1em;
   font-size: 28px !important;
