+++ app/src/pages/Login/EmailLogin.js
@@ -10,8 +10,8 @@ import {
   withState
 } from 'recompose'
 import { graphql } from 'react-apollo'
-import gql from 'graphql-tag'
 import { Form } from 'react-form'
+import { emailLoginGQL } from './LoginGQL'
 import ReactFormInput from '../../components/react-form-semantic-ui-react/ReactFormInput'
 import './EmailLogin.css'
 const validate = require('validate.js')
@@ -179,14 +179,6 @@ const EmailLogin = ({
     </div>
   )
 }
-
-const emailLoginGQL = gql`
-  mutation emailLogin($email: String!, $username: String!) {
-    emailLogin(email: $email, username: $username) {
-      success
-    }
-  }
-`
 
 export default compose(
   withState('isPending', 'onPending', false),
