+++ app/src/pages/Login/EthLogin.js
@@ -3,12 +3,12 @@ import { connect } from 'react-redux'
 import { graphql, withApollo } from 'react-apollo'
 import GoogleAnalytics from 'react-ga'
 import Raven from 'raven-js'
-import gql from 'graphql-tag'
 import {
   lifecycle,
   compose
 } from 'recompose'
 import { Message } from 'semantic-ui-react'
+import { ethLoginGQL, followedProjectsGQL } from './LoginGQL'
 import {
   setupWeb3,
   hasMetamask,
@@ -23,6 +23,7 @@ const EthLogin = ({
   requestAuth,
   checkMetamask,
   authWithSAN,
+  followedProjects,
   client
 }) => {
   return (
@@ -49,7 +50,7 @@ const EthLogin = ({
           account={user.account}
           pending={user.isLoading}
           error={user.error}
-          handleAuth={() => requestAuth(user.account, authWithSAN, client)} />}
+          handleAuth={() => requestAuth(user.account, authWithSAN, followedProjects, client)} />}
     </Fragment>
   )
 }
@@ -68,7 +69,7 @@ const mapDispatchToProps = dispatch => {
         hasMetamask
       })
     },
-    requestAuth: (address, authWithSAN, client) => {
+    requestAuth: (address, authWithSAN, followedProjects, client) => {
       signMessage(address).then(({messageHash, signature}) => {
         dispatch({
           type: 'PENDING_LOGIN'
@@ -87,7 +88,16 @@ const mapDispatchToProps = dispatch => {
             user
           })
           client.resetStore()
-        }).catch((error) => {
+        })
+        .then(() => {
+          followedProjects.refetch().then(({data: {followedProjects}}) => {
+            dispatch({
+              type: 'SET_FOLLOWED_PROJECTS',
+              followedProjects
+            })
+          })
+        })
+        .catch((error) => {
           dispatch({
             type: 'FAILED_LOGIN',
             errorMessage: error
@@ -118,25 +128,6 @@ const mapDispatchToProps = dispatch => {
   }
 }
 
-const ethLoginGQL = gql`
-  mutation ethLogin($signature: String!, $address: String!, $messageHash: String!) {
-    ethLogin(
-      signature: $signature,
-      address: $address,
-      messageHash: $messageHash) {
-        token,
-        user {
-          id,
-          email,
-          username,
-          ethAccounts {
-            address,
-            sanBalance
-          }
-        }
-      }
-}`
-
 export default compose(
   connect(
     mapStateToProps,
@@ -145,6 +136,9 @@ export default compose(
   withApollo,
   graphql(ethLoginGQL, {
     name: 'authWithSAN'
+  }),
+  graphql(followedProjectsGQL, {
+    name: 'followedProjects'
   }),
   lifecycle({
     componentDidMount () {
